# openapi.yml
openapi: 3.0.3
info:
  title: Apple Reminder Web App API
  description: Apple Reminder 웹 앱 백엔드 API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Lists
    description: 리마인더 리스트 관리
  - name: Reminders
    description: 리마인더 관리
  - name: Smart Lists
    description: 스마트 리스트 (오늘/예정/전체/플래그/완료됨)

paths:
  /api/lists:
    get:
      tags: [Lists]
      summary: 전체 리스트 조회
      operationId: findAllLists
      responses:
        '200':
          description: 리스트 목록 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReminderList'

    post:
      tags: [Lists]
      summary: 리스트 생성
      operationId: createList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderListCreate'
      responses:
        '201':
          description: 리스트 생성 완료
          headers:
            Location:
              description: 생성된 리스트의 URI
              schema:
                type: string
                example: /api/lists/4
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderList'

  /api/lists/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: 리스트 ID
        schema:
          type: integer
          format: int64

    get:
      tags: [Lists]
      summary: 리스트 단건 조회
      operationId: findListById
      responses:
        '200':
          description: 리스트 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderList'
        '404':
          description: 리스트를 찾을 수 없음

    put:
      tags: [Lists]
      summary: 리스트 수정
      operationId: updateList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderListUpdate'
      responses:
        '200':
          description: 수정된 리스트 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderList'

    delete:
      tags: [Lists]
      summary: 리스트 삭제
      operationId: deleteList
      responses:
        '204':
          description: 삭제 완료 (응답 본문 없음)

  /api/reminders:
    get:
      tags: [Reminders]
      summary: 리스트별 리마인더 조회
      operationId: findRemindersByListId
      parameters:
        - name: listId
          in: query
          required: true
          description: 리스트 ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 리마인더 목록 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'

    post:
      tags: [Reminders]
      summary: 리마인더 생성
      operationId: createReminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderCreate'
      responses:
        '201':
          description: 리마인더 생성 완료
          headers:
            Location:
              description: 생성된 리마인더의 URI
              schema:
                type: string
                example: /api/reminders/10
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'

  /api/reminders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: 리마인더 ID
        schema:
          type: integer
          format: int64

    get:
      tags: [Reminders]
      summary: 리마인더 단건 조회
      operationId: findReminderById
      responses:
        '200':
          description: 리마인더 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '404':
          description: 리마인더를 찾을 수 없음

    put:
      tags: [Reminders]
      summary: 리마인더 수정
      operationId: updateReminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderUpdate'
      responses:
        '200':
          description: 수정된 리마인더 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'

    delete:
      tags: [Reminders]
      summary: 리마인더 삭제
      operationId: deleteReminder
      responses:
        '204':
          description: 삭제 완료 (응답 본문 없음)

  /api/reminders/{id}/toggle:
    parameters:
      - name: id
        in: path
        required: true
        description: 리마인더 ID
        schema:
          type: integer
          format: int64
    patch:
      tags: [Reminders]
      summary: 리마인더 완료 토글
      operationId: toggleReminderComplete
      responses:
        '200':
          description: 토글된 리마인더 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'

  /api/reminders/today:
    get:
      tags: [Smart Lists]
      summary: 오늘 마감 리마인더 조회
      operationId: findTodayReminders
      responses:
        '200':
          description: 오늘 마감 리마인더 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'

  /api/reminders/scheduled:
    get:
      tags: [Smart Lists]
      summary: 예정된 리마인더 조회
      operationId: findScheduledReminders
      responses:
        '200':
          description: 마감일이 설정된 미완료 리마인더 목록 (날짜순)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'

  /api/reminders/all:
    get:
      tags: [Smart Lists]
      summary: 전체 미완료 리마인더 조회
      operationId: findAllIncompleteReminders
      responses:
        '200':
          description: 모든 미완료 리마인더 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'

  /api/reminders/flagged:
    get:
      tags: [Smart Lists]
      summary: 플래그 리마인더 조회
      operationId: findFlaggedReminders
      responses:
        '200':
          description: 플래그가 설정된 미완료 리마인더 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'

  /api/reminders/completed:
    get:
      tags: [Smart Lists]
      summary: 완료된 리마인더 조회
      operationId: findCompletedReminders
      responses:
        '200':
          description: 완료된 리마인더 목록 (완료일시 내림차순)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'

  /api/reminders/search:
    get:
      tags: [Reminders]
      summary: 리마인더 검색
      operationId: searchReminders
      parameters:
        - name: q
          in: query
          required: true
          description: 검색 키워드 (제목, 메모)
          schema:
            type: string
      responses:
        '200':
          description: 검색 결과 리마인더 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'

  /api/reminders/counts:
    get:
      tags: [Smart Lists]
      summary: 스마트 리스트 카운트 조회
      operationId: getReminderCounts
      responses:
        '200':
          description: 각 스마트 리스트별 카운트
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartListCounts'

components:
  schemas:
    ReminderList:
      type: object
      description: 리마인더 리스트
      properties:
        id:
          type: integer
          format: int64
          description: 리스트 ID
          example: 1
        name:
          type: string
          description: 리스트 이름
          example: 업무
        color:
          type: string
          description: 리스트 색상 코드
          example: '#007AFF'
        icon:
          type: string
          nullable: true
          description: 리스트 아이콘 이름
          example: briefcase
        sortOrder:
          type: integer
          description: 정렬 순서
          example: 1
        createdAt:
          type: string
          format: date-time
          description: 생성일시
        updatedAt:
          type: string
          format: date-time
          description: 수정일시

    ReminderListCreate:
      type: object
      description: 리스트 생성 요청
      required:
        - name
      properties:
        name:
          type: string
          description: 리스트 이름
          example: 새 리스트
        color:
          type: string
          description: 리스트 색상 코드
          default: '#007AFF'
          example: '#FF3B30'
        icon:
          type: string
          nullable: true
          description: 리스트 아이콘 이름
          example: star
        sortOrder:
          type: integer
          description: 정렬 순서
          default: 0
          example: 4

    Reminder:
      type: object
      description: 리마인더
      properties:
        id:
          type: integer
          format: int64
          example: 1
        listId:
          type: integer
          format: int64
          example: 2
        title:
          type: string
          example: 프로젝트 회의 준비
        memo:
          type: string
          nullable: true
          example: 발표자료 정리
        dueDate:
          type: string
          format: date-time
          nullable: true
        priority:
          type: integer
          description: 0:없음, 1:낮음, 2:보통, 3:높음
          example: 2
        isFlagged:
          type: boolean
          example: false
        isCompleted:
          type: boolean
          example: false
        completedAt:
          type: string
          format: date-time
          nullable: true
        sortOrder:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ReminderCreate:
      type: object
      description: 리마인더 생성 요청
      required:
        - listId
        - title
      properties:
        listId:
          type: integer
          format: int64
          example: 2
        title:
          type: string
          example: 새로운 할 일
        memo:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        priority:
          type: integer
          default: 0
          example: 1
        isFlagged:
          type: boolean
          default: false
        sortOrder:
          type: integer
          default: 0

    ReminderUpdate:
      type: object
      description: 리마인더 수정 요청
      properties:
        listId:
          type: integer
          format: int64
        title:
          type: string
        memo:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        priority:
          type: integer
        isFlagged:
          type: boolean
        sortOrder:
          type: integer

    SmartListCounts:
      type: object
      description: 스마트 리스트별 카운트
      properties:
        today:
          type: integer
          description: 오늘 마감 리마인더 수
          example: 3
        scheduled:
          type: integer
          description: 예정된 리마인더 수
          example: 5
        total:
          type: integer
          description: 전체 미완료 리마인더 수
          example: 12
        flagged:
          type: integer
          description: 플래그 리마인더 수
          example: 2
        completed:
          type: integer
          description: 완료된 리마인더 수
          example: 8

    ReminderListUpdate:
      type: object
      description: 리스트 수정 요청
      properties:
        name:
          type: string
          description: 리스트 이름
          example: 수정된 리스트
        color:
          type: string
          description: 리스트 색상 코드
          example: '#34C759'
        icon:
          type: string
          nullable: true
          description: 리스트 아이콘 이름
          example: folder
        sortOrder:
          type: integer
          description: 정렬 순서
          example: 2